Concordance test for binomial models

Performs a concordance/disconcordance (C-statistic) test on binomial models.

concordance(y, p)

Arguments

y
vector of binomial response variable used in model

p
estimated probabilities from fit binomial model

Value

list object with: concordance, discordance, tied and pairs

Note

Test of binomial regression for the hypothesis that probabilities of all positives [1], are greater than the probabilities of the nulls [0]. The concordance would be 100 inverse of concordance, representing the null. The C-statistic has been show to be comparable to the area under an ROC

Results are: concordance - percent of positives that are greater than probabilities of nulls. discordance - concordance inverse of concordance representing the null class, tied - number of tied probabilities and pairs - number of pairs compared

References

Austin, P.C. & E.W. Steyerberg (2012) Interpreting the concordance statistic of a logistic regression model: relation to the variance and odds ratio of a continuous explanatory variable. BMC Medical Research Methodology, 12:82

Harrell, F.E. (2001) Regression modelling strategies. Springer, New York, NY.

Royston, P. & D.G. Altman (2010) Visualizing and assessing discrimination in the logistic regression model. Statistics in Medicine 29(24):2508-2520

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org>

Examples

data(mtcars)
dat <- subset(mtcars, select=c(mpg, am, vs))
glm.reg <- glm(vs ~ mpg, data = dat, family = binomial)
concordance(dat$vs, predict(glm.reg, type = "response")) 

#> $concordance
#> [1] 0.9087302
#> 
#> $discordance
#> [1] 0.09126984
#> 
#> $tied
#> [1] 1.387779e-17
#> 
#> $pairs
#> [1] 252
#>