Cosine Similarity Index

Calculates the cosine similarity and angular similarity on two vectors or a matrix

csi(x, y = NULL)

Arguments

x
A vector or matrix object

y
If x is a vector, then a vector object

Value

If x is a matrix, a list object with: similarity and angular.similarity matrices or, if x and y are vectors, a vector of similarity and angular.similarity

Note

The cosine similarity index is a measure of similarity between two vectors of an inner product space. This index is bested suited for high-dimensional positive variable space. One useful application of the index is to measure separability of clusters derived from algorithmic approaches (e.g., k-means). It is a good common practice to center the data before calculating the index. It should be noted that the cosine similarity index is mathematically, and often numerically, equivalent to the Pearson's correlation coefficient

The cosine similarity index is derived: s(xy) = x * y / ||x|| * ||y||, where the expected is 1.0 (perfect similarity) to -1.0 (perfect dissimilarity). A normalized angle between the vectors can be used as a bounded similarity function within [0,1] angular similarity = 1 - (cos(s)^-1/pi)

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org>

Examples

# Compare two vectors (centered using scale)
  x=runif(100)
  y=runif(100)^2
  csi(as.vector(scale(x)),as.vector(scale(y)))

#>         similarity angular.similarity 
#>         0.09173184         0.68034616 

  
  #' # Compare columns (vectors) in a matrix (centered using scale)
  x <- matrix(round(runif(100),0),nrow=20,ncol=5)
  ( s <- csi(scale(x)) )

#> $similarity
#>             [,1]          [,2]        [,3]          [,4]        [,5]
#> [1,]  1.00000000  6.579517e-02  0.21212121 -1.005038e-01  0.03160698
#> [2,]  0.06579517  1.000000e+00 -0.06579517 -1.168656e-17  0.25162996
#> [3,]  0.21212121 -6.579517e-02  1.00000000 -1.005038e-01 -0.03160698
#> [4,] -0.10050378 -1.168656e-17 -0.10050378  1.000000e+00  0.52414242
#> [5,]  0.03160698  2.516300e-01 -0.03160698  5.241424e-01  1.00000000
#> 
#> $angular.similarity
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 1.0000000 0.6809999 0.6743921 0.6800757 0.6815311
#> [2,] 0.6809999 1.0000000 0.6809999 0.6816901 0.6713399
#> [3,] 0.6743921 0.6809999 1.0000000 0.6800757 0.6815311
#> [4,] 0.6800757 0.6816901 0.6800757 1.0000000 0.6323319
#> [5,] 0.6815311 0.6713399 0.6815311 0.6323319 1.0000000
#> 

    
# Compare vector (x) to each column in a matrix (y)
y <- matrix(round(runif(500),3),nrow=100,ncol=5)
x=runif(100) 
csi(as.vector(scale(x)),scale(y))

#>                          [,1]        [,2]       [,3]       [,4]       [,5]
#> similarity         0.04670747 -0.08636181 0.03697387 -0.1701165 0.02584444
#> angular.similarity 0.68134259  0.68049938 0.68147241  0.6770280 0.68158378