Insert a row or column into a data.frame

Inserts a new row or column into a data.frame at a specified location

insert(x, MARGIN = 1, value = NULL, idx, name = NULL)

Arguments

x
Existing data.frame

MARGIN
Insert a 1 = row or 2 = column

value
A vector of values equal to the length of MARGIN, if nothing specified values with be NA

idx
Index position to insert row or column

name
Name of new column (not used for rows, MARGIN=1)

Value

A data.frame with the new row or column inserted

Note

Where there are methods to easily add a row/column to the end or beginning of a data.frame, it is not straight forward to insert data at a specific location within the data.frame. This function allows for inserting a vector at a specific location eg., between columns or rows 1 and 2 where row/column 2 is moved to the 3rd position and a new vector of values is inserted into the 2nd position.

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org>

Examples

d <- data.frame(ID=1:10, y=runif(10))

# insert row
insert(d, idx=2)

#> Inserting row 

#>    ID          y
#> 1   1 0.68355056
#> 2  NA         NA
#> 3   2 0.97202445
#> 4   3 0.07438633
#> 5   4 0.72655659
#> 6   5 0.49678935
#> 7   6 0.55303877
#> 8   7 0.03892912
#> 9   8 0.67843460
#> 10  9 0.06465576
#> 11 10 0.74824150

insert(d, value=c(20,0), idx=2)

#> Inserting row 

#>    ID          y
#> 1   1 0.68355056
#> 2  20 0.00000000
#> 3   2 0.97202445
#> 4   3 0.07438633
#> 5   4 0.72655659
#> 6   5 0.49678935
#> 7   6 0.55303877
#> 8   7 0.03892912
#> 9   8 0.67843460
#> 10  9 0.06465576
#> 11 10 0.74824150

# insert column
insert(d, MARGIN=2, idx=2)

#> Inserting column 

#>    ID V1          y
#> 1   1 NA 0.68355056
#> 2   2 NA 0.97202445
#> 3   3 NA 0.07438633
#> 4   4 NA 0.72655659
#> 5   5 NA 0.49678935
#> 6   6 NA 0.55303877
#> 7   7 NA 0.03892912
#> 8   8 NA 0.67843460
#> 9   9 NA 0.06465576
#> 10 10 NA 0.74824150

insert(d, MARGIN = 2, value = rep(0,10), idx=2, name="x")

#> Inserting column 

#>    ID x          y
#> 1   1 0 0.68355056
#> 2   2 0 0.97202445
#> 3   3 0 0.07438633
#> 4   4 0 0.72655659
#> 5   5 0 0.49678935
#> 6   6 0 0.55303877
#> 7   7 0 0.03892912
#> 8   8 0 0.67843460
#> 9   9 0 0.06465576
#> 10 10 0 0.74824150