Logistic and Auto-logistic regression

Performs a logistic (binomial) or auto-logistic (spatially lagged binomial) regression using maximum likelihood or penalized maximum likelihood estimation.

It should be noted that the auto-logistic model (Besag 1972) is intended for exploratory analysis of spatial effects. Auto-logistic are know to underestimate the effect of environmental variables and tend to be unreliable (Dormann 2007). Wij matrix options under style argument - B is the basic binary coding, W is row standardized (sums over all links to n), C is globally standardized (sums over all links to n), U is equal to C divided by the number of neighbours (sums over all links to unity) and S is variance-stabilizing. Spatially lagged y defined as: W(y)ij=sumj_(Wij yj)/ sumj_(Wij) where; Wij=1/Euclidean(i,j) If the object passed to the function is an sp class there is no need to call the data slot directly via "object@data", just pass the object name.

logistic.regression(
  ldata,
  y,
  x,
  penalty = TRUE,
  autologistic = FALSE,
  coords = NULL,
  bw = NULL,
  type = "inverse",
  style = "W",
  longlat = FALSE,
  ...
)

Arguments

ldata
data.frame object containing variables

y
Dependent variable (y) in ldata

x
Independent variable(s) (x) in ldata

penalty
Apply regression penalty (TRUE/FALSE)

autologistic
Add auto-logistic term (TRUE/FALSE)

coords
Geographic coordinates for auto-logistic model matrix or sp object.

bw
Distance bandwidth to calculate spatial lags (if empty neighbors result, need to increase bandwidth). If not provided it will be calculated automatically based on the minimum distance that includes at least one neighbor.

type
Neighbor weighting scheme (see autocov_dist)

style
Type of neighbor matrix (Wij), default is mean of neighbors

longlat
Are coordinates (coords) in geographic, lat/long (TRUE/FALSE)

...
Additional arguments passed to lrm

Value

A list class object with the following components:

model - lrm model object (rms class)

bandwidth - If AutoCov = TRUE returns the distance bandwidth used for the auto-covariance function

diagTable - data.frame of regression diagnostics

coefTable - data.frame of regression coefficients

Residuals - data.frame of residuals and standardized residuals

AutoCov - If an auto-logistic model, AutoCov represents lagged auto-covariance term

References

Besag, J.E., (1972) Nearest-neighbour systems and the auto-logistic model for binary data. Journal of the Royal Statistical Society, Series B Methodological 34:75-83

Dormann, C.F., (2007) Assessing the validity of autologistic regression. Ecological Modelling 207:234-242

Le Cessie, S., Van Houwelingen, J.C., (1992) Ridge estimators in logistic regression. Applied Statistics 41:191-201

Shao, J., (1993) Linear model selection by cross-validation. JASA 88:486-494

See also

Author

Jeffrey S. Evans jeffrey_evans@tnc.org

Examples

require(sp)
require(spdep)
require(rms)                                                                       

#> Loading required package: rms

#> Warning: there is no package called 'rms'

data(meuse)
  coordinates(meuse) <- ~x+y  
    meuse@data <- data.frame(DepVar=rbinom(dim(meuse)[1], 1, 0.5), 
                              meuse@data)

#### Logistic model
lmodel <- logistic.regression(meuse, y='DepVar', 
                  x=c('dist','cadmium','copper')) 

#> Error in logistic.regression(meuse, y = "DepVar", x = c("dist", "cadmium",     "copper")): please install rms package before running this function

  lmodel$model

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

    lmodel$diagTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

      lmodel$coefTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

#### Logistic model with factorial variable
lmodel <- logistic.regression(meuse, y='DepVar', 
            x=c('dist','cadmium','copper', 'soil')) 

#> Error in logistic.regression(meuse, y = "DepVar", x = c("dist", "cadmium",     "copper", "soil")): please install rms package before running this function

  lmodel$model

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

    lmodel$diagTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

      lmodel$coefTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

### Auto-logistic model using 'autocov_dist' in 'spdep' package
lmodel <- logistic.regression(meuse, y='DepVar', 
            x=c('dist','cadmium','copper'), autologistic=TRUE, 
            coords=coordinates(meuse), bw=5000) 

#> Error in logistic.regression(meuse, y = "DepVar", x = c("dist", "cadmium",     "copper"), autologistic = TRUE, coords = coordinates(meuse),     bw = 5000): please install rms package before running this function

  lmodel$model

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

    lmodel$diagTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

      lmodel$coefTable

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

  est <- predict(lmodel$model, type='fitted.ind')

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'predict': object 'lmodel' not found

#### Add residuals, standardized residuals and estimated probabilities
VarNames <- rownames(lmodel$model$var)[-1]

#> Error in rownames(lmodel$model$var): object 'lmodel' not found

  meuse@data$AutoCov <- lmodel$AutoCov

#> Error in eval(expr, envir, enclos): object 'lmodel' not found

    meuse@data <- data.frame(meuse@data, Residual=lmodel$Residuals[,1], 
                     StdResid=lmodel$Residuals[,2], Probs=predict(lmodel$model, 
                     meuse@data[,VarNames],type='fitted') )  

#> Error in data.frame(meuse@data, Residual = lmodel$Residuals[, 1], StdResid = lmodel$Residuals[,     2], Probs = predict(lmodel$model, meuse@data[, VarNames],     type = "fitted")): object 'lmodel' not found

#### Plot fit and probabilities
resid(lmodel$model, "partial", pl="loess") 

#> Error in resid(lmodel$model, "partial", pl = "loess"): object 'lmodel' not found

# plot residuals
resid(lmodel$model, "partial", pl=TRUE) 

#> Error in resid(lmodel$model, "partial", pl = TRUE): object 'lmodel' not found

# global test of goodness of fit 
resid(lmodel$model, "gof")

#> Error in resid(lmodel$model, "gof"): object 'lmodel' not found

# Approx. leave-out linear predictors
lp1 <- resid(lmodel$model, "lp1")            

#> Error in resid(lmodel$model, "lp1"): object 'lmodel' not found

# Approx leave-out-1 deviance            
-2 * sum(meuse@data$DepVar * lp1 + log(1-plogis(lp1)))

#> Error in eval(expr, envir, enclos): object 'lp1' not found

# plot estimated probabilities at points
spplot(meuse, c('Probs'))

#> Error in create.z(as(obj, "data.frame"), zcol): no support for variable of this type