Average Nearest Neighbor Index (NNI)

Calculates the NNI as a measure of clustering or dispersal

The nearest neighbor index is expressed as the ratio of the observed distance divided by the expected distance. The expected distance is the average distance between neighbors in a hypothetical random distribution. If the index is less than 1, the pattern exhibits clustering; if the index is greater than 1, the trend is toward dispersion or competition. The Nearest Neighbor Index is calculated as:

Mean Nearest Neighbor Distance (observed) D(nn) = sum(min(Dij)/N)

Mean Random Distance (expected) D(e) = 0.5 SQRT(A/N)

Nearest Neighbor Index NNI = D(nn)/D(e) Where; D=neighbor distance, A=Area

nni(x, win = "hull")

Arguments

x
An sp point object

win
Type of window 'hull' or 'extent'

Value

list object containing NNI = nearest neighbor index, z.score = Z Score value, p = p value, expected.mean.distance = Expected mean distance, observed.mean.distance = Observed meand distance.

References

Clark, P.J., and F.C. Evans (1954) Distance to nearest neighbour as a measure of spatial relationships in populations. Ecology 35:445-453

Cressie, N (1991) Statistics for spatial data. Wiley & Sons, New York.

Author

Jeffrey S. Evans jeffrey_evans@tnc.org

Examples

require(sp)
data(meuse)
  coordinates(meuse) <- ~x+y
    nni(meuse)

#> $NNI
#> [1] 1.194171
#> 
#> $z.score
#> [1] 4.624676
#> 
#> $p
#> [1] 3.75184e-06
#> 
#> $expected.mean.distance
#> [1] 93.52889
#> 
#> $observed.mean.distance
#> [1] 111.6895
#>