Biodiversity Planning Units

Subset of biodiversity planning units for Haiti ecoregional spatial reserve plan

Format

A sp SpatialPolygonsDataFrame with 5919 rows and 46 variables:

UNIT_ID

Unique planning unit ID

DR_Dr_A

Biodiversity target

DR_Dr_L

Biodiversity target

Ht_Dr_A

Biodiversity target

Ht_Dr_L

Biodiversity target

DR_Ms_A

Biodiversity target

DR_Ms_L

Biodiversity target

Ht_Ms_L

Biodiversity target

DR_LM_M

Biodiversity target

H_LM_M_L

Biodiversity target

H_LM_R_L

Biodiversity target

DR_LM_R_L

Biodiversity target

DR_Rn_L

Biodiversity target

DR_LM_R_S

Biodiversity target

DR_Rn_S

Biodiversity target

DR_Ms_S

Biodiversity target

Ht_Ms_A

Biodiversity target

DR_Ms_E

Biodiversity target

DR_Ms_I

Biodiversity target

DR_Rn_E

Biodiversity target

DR_Rn_I

Biodiversity target

H_LM_R_E

Biodiversity target

Ht_Ms_E

Biodiversity target

Ht_Rn_E

Biodiversity target

DR_Rn_A

Biodiversity target

Ht_Rn_A

Biodiversity target

Ht_Rn_I

Biodiversity target

Ht_Dr_E

Biodiversity target

Ht_Ms_S

Biodiversity target

Ht_Dr_S

Biodiversity target

Ht_Rn_L

Biodiversity target

Ht_Th_A

Biodiversity target

Ht_Th_L

Biodiversity target

Ht_Th_S

Biodiversity target

Ht_Dr_U

Biodiversity target

Ht_Dr_I

Biodiversity target

Ht_Ms_I

Biodiversity target

H_LM_M_A

Biodiversity target

H_LM_M_E

Biodiversity target

H_LM_R_A

Biodiversity target

H_LM_M_S

Biodiversity target

H_LM_R_I

Biodiversity target

H_LM_R_S

Biodiversity target

Ht_Rn_S

Biodiversity target

Ht_Ms_U

Biodiversity target

Ht_Rn_U

Biodiversity target

Source

https://www.conservationgateway.org

References

Evans, J.S., S.R. Schill, G.T. Raber (2015) A Systematic Framework for Spatial Conservation Planning and Ecological Priority Design in St. Lucia, Eastern Caribbean. Chapter 26 in Central American Biodiversity : Conservation, Ecology and a Sustainable Future. F. Huettman (eds). Springer, NY.