spectral separability

Calculates spectral separability by class

Available statistics:

Bhattacharyya distance (Bhattacharyya 1943; Harold 2003) measures the similarity of two discrete or continuous probability distributions.

Jeffries-Matusita (default) distance (Bruzzone et al., 2005; Swain et al., 1971) is a scaled (0-2) version of Bhattacharyya. The J-M distance is asymptotic to 2, where 2 suggest complete separability.

spectral.separability(x, y, jeffries.matusita = TRUE)

Arguments

x
data.frame, matrix or vector of spectral values must, match classes defined in y

y
A vector or factor with grouping classes, must match row wise values in x

jeffries.matusita
(TRUE/FALSE) Return J-M distance (default) else Bhattacharyya

Value

A matrix of class-wise Jeffries-Matusita or Bhattacharyya distance separability values

References

Bhattacharyya, A. (1943) On a measure of divergence between two statistical populations defined by their probability distributions'. Bulletin of the Calcutta Mathematical Society 35:99-109

Bruzzone, L., F. Roli, S.B. Serpico (1995) An extension to multiclass cases of the Jefferys-Matusita distance. IEEE Transactions on Pattern Analysis and Machine Intelligence 33:1318-1321

Kailath, T., (1967) The Divergence and Bhattacharyya measures in signal selection. IEEE Transactions on Communication Theory 15:52-60

Author

Jeffrey S. Evans jeffrey_evans@tnc.org

Examples

#' # Create example data 
require(MASS)                

#> Loading required package: MASS

#> 
#> Attaching package: 'MASS'

#> The following object is masked from 'package:spatstat.geom':
#> 
#>     area

#> The following objects are masked from 'package:raster':
#> 
#>     area, select

d <- 6                 # Number of bands
n.class <- 5           # Number of classes
n <- rep(1000, 5)                
mu <- round(matrix(rnorm(d*n.class, 128, 1), 
            ncol=n.class, byrow=TRUE), 0)

x <- matrix(double(), ncol=d, nrow=0)
  classes <- integer()
    for (i in 1:n.class) {
      f <- svd(matrix(rnorm(d^2), ncol=d))
      sigma <- t(f$v) %*% diag(rep(10, d)) %*% f$v
      x <- rbind(x, mvrnorm(n[i], mu[, i], sigma))
      classes <- c(classes, rep(i, n[i]))
    }
colnames(x) <- paste0("band", 1:6)
classes <- factor(classes, labels=c("water", "forest",
                  "shrub", "urban", "ag"))

# Separability for multi-band (multivariate) spectra 
spectral.separability(x, classes)

#> Calculating distances for 5 classes and 6 spectra

#>            water    forest     shrub     urban        ag
#> water  0.0000000 0.3778270 0.7143928 0.9339081 0.7732728
#> forest 0.3778270 0.0000000 0.6163500 0.7661204 0.6390126
#> shrub  0.7143928 0.6163500 0.0000000 0.7409606 0.6064753
#> urban  0.9339081 0.7661204 0.7409606 0.0000000 0.4719211
#> ag     0.7732728 0.6390126 0.6064753 0.4719211 0.0000000

# Separability for single-band (univariate) spectra 
spectral.separability(x[,1], classes)

#> Calculating distances for 5 classes and 1 spectra

#>            water     forest     shrub      urban         ag
#> water  0.0000000 0.17635922 0.2864637 0.17968714 0.14952478
#> forest 0.1763592 0.00000000 0.4539584 0.01049628 0.04044985
#> shrub  0.2864637 0.45395839 0.0000000 0.45504275 0.42085442
#> urban  0.1796871 0.01049628 0.4550427 0.00000000 0.03703873
#> ag     0.1495248 0.04044985 0.4208544 0.03703873 0.00000000