Time to event

Returns the time (sum to position) to a specified value

The time to event represents the sum of positions, in the vector, until the specified value is found ie., (0,0,1) would be 3 or, 2 with up.to=TRUE. The int argument allows for rounding a continuous variable. Since it may be difficult to fine an exact match to a floating point value rounding mitigates the problem. If you want a specific rounding value (eg., 1 decimal place) you can apply it to x first then pass it to the function.

time_to_event(x, y = 1, dir = c("LR", "RL"), int = FALSE, up.to = FALSE)

Arguments

x
A vector, representing time-series, to evaluate

y
Threshold value tor return position for

dir
Direction of evaluation c("LR", "RL")

int
FALSE | TRUE - Evaluate as integer (rounds to 0 decimal places)

up.to
FALSE | TRUE - Return value before event

Value

A vector value representing the time to event

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org>

Examples

library(raster)

# Binomial instance
time_to_event(c(0,0,0,0,1,0,0,0,1,0))

#> [1] 5

time_to_event(c(0,0,0,0,1,0,0,0,1,0), up.to = TRUE)

#> [1] 4

time_to_event(c(0,0,0,0,1,0,0,0,1,0), dir="RL")

#> [1] 2

r <- do.call(raster::stack, replicate(20,raster::raster(matrix(sample(
             c(0,1), 1000, replace=TRUE), 100, 100))))             
  ( t2e <- calc(r, fun=time_to_event) )

#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 0.01, 0.01  (x, y)
#> extent     : 0, 1, 0, 1  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 2, 13  (min, max)
#> 

# Continuous threshold instance
( x <- runif(100, 0,7) ) 

#>   [1] 1.07396994 5.70921598 5.55369217 1.30501700 4.58921485 3.22395906
#>   [7] 0.67182842 4.52615698 1.95761455 6.64697271 4.53199079 0.47777812
#>  [13] 6.47031219 4.14591552 3.89770455 0.78113859 1.68307554 2.23647627
#>  [19] 3.80515944 6.87421845 3.78753266 2.29460205 0.95003337 4.80559262
#>  [25] 1.28342069 6.04114354 6.32295695 5.44153554 1.37412579 1.14233735
#>  [31] 3.77863480 0.63453013 5.44423631 0.65680388 0.64408868 0.86140983
#>  [37] 1.13985652 3.35195038 3.95084955 5.67568747 5.00147117 4.62782958
#>  [43] 6.94450252 0.17641038 6.22974357 1.81386573 5.17411717 6.11439222
#>  [49] 1.64346204 0.41451328 0.17077338 6.35674847 5.75852600 3.79772506
#>  [55] 5.41608223 4.20062447 1.49360474 6.60480713 2.45339776 5.58663507
#>  [61] 4.42058480 5.53521921 2.25281942 5.78511271 4.08347124 3.41045610
#>  [67] 3.42510275 2.98933848 0.02570184 5.45582732 6.05942256 6.16285084
#>  [73] 0.26911113 4.65030289 6.33381557 2.06937244 0.36090119 4.78668165
#>  [79] 4.99216313 3.42889945 6.55607370 6.88947469 6.22435794 2.39618316
#>  [85] 0.72807553 0.90011868 5.78320981 1.40506663 6.14158755 3.42067389
#>  [91] 5.42493981 2.04106386 4.62411418 6.25058059 5.87053715 0.25641993
#>  [97] 1.64831922 2.51730922 3.85065175 6.94500737

time_to_event(x, y = 5, int=TRUE)

#> [1] 2

r <- do.call(raster::stack, replicate(20,raster::raster(matrix(
             runif(1000,0,7), 100, 100))))
  t2e <- function(x) { time_to_event(x, y=5, int=TRUE) }              
  ( t2e <- calc(r, fun=time_to_event) )

#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 0.01, 0.01  (x, y)
#> extent     : 0, 1, 0, 1  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 2, 5  (min, max)
#>